% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FoliometriK.R
\name{run_FoliometriK}
\alias{run_FoliometriK}
\title{Run \emph{\strong{FoliometriK}} Application}
\usage{
run_FoliometriK()
}
\description{
This function runs the \emph{\strong{FoliometriK}} application in the user's default web browser.
}
\section{Authors}{

Victor M. Ramirez-Arrieta and Dennis Denis-Avila
}

\examples{
run_FoliometriK()
}
